8_OpenAI_Agent_SDK/pyproject.toml
[project]
name = "8-openai-agent-sdk"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.12"

dependencies = [
    "openai (>=1.77.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "openai-agents (>=0.0.14,<0.0.15)",
    "regex (>=2024.11.6,<2025.0.0)",
    "streamlit (>=1.45.0,<2.0.0)"

]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
